集成支付宝支付（半成品）

+++ src/main/java/com/csk/mmall/service/IOrderService.java
@@ -3,5 +3,5 @@ package com.csk.mmall.service;
 import com.csk.mmall.common.ServerResponse;
 
 public interface IOrderService {
-    ServerResponse pay(Integer orderId, Integer id, String path);
+    ServerResponse pay(Long orderId, Integer id, String path);
 }
